+++ src/SignalR/clients/cpp/include/signalrclient/http_client.h
