+++ src/SignalR/clients/cpp/include/signalrclient/transfer_format.h
