+++ src/SignalR/clients/cpp/include/signalrclient/websocket_client.h
